+++ packages/core/__tests__/utils.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import { Document, Source } from '../src/resolve';
 import { NormalizedProblem } from '../src/walk';
 import { RuleConfig, LintConfig, Plugin } from '../src/config/config';
-import { Oas3RuleSet } from '../src/lint';
+import { Oas3RuleSet } from '../src/oas-types';
 
 export function parseYamlToDocument(body: string, absoluteRef: string = ''): Document {
   return {
