+++ packages/core/src/bundle.ts
@@ -6,9 +6,9 @@ import { Oas2Types } from './types/oas2';
 import { Oas3_1Types } from './types/oas3_1';
 import { NormalizedNodeType, normalizeTypes, NodeType } from './types';
 import { WalkContext, walkDocument, UserContext, ResolveResult } from './walk';
-import { detectOpenAPI, openAPIMajor, OasMajorVersion } from './lint';
+import { detectOpenAPI, openAPIMajor, OasMajorVersion } from './oas-types';
 import { Location, refBaseName } from './ref-utils';
-import { Config, LintConfig } from './config/config';
+import type { Config, LintConfig } from './config/config';
 import { initRules } from './config/rules';
 import { reportUnresolvedRef } from './rules/no-unresolved-refs';
 import { isPlainObject } from './utils';
